Resolved conflicts

+++ Iteach/IteachAPI/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -128,11 +128,11 @@ namespace IteachAPI.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<int>("TestId")
-                        .HasColumnType("int");
+                    b.Property<long>("TestId")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("UserId")
-                        .HasColumnType("int");
+                    b.Property<long>("UserId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
